move wrapMorphismWithCheck to the Morphism module
